[svn-r12767] Description:     Code cleanup to improve formatting & reduce compiler warnings.
Tested on:
    Linux/32 2.6 (chicago)
    Linux/64 2.6 (chicago2)

+++ src/H5S.c
@@ -1720,7 +1720,7 @@ H5S_encode(H5S_t *obj, unsigned char *buf, size_t *nalloc)
     FUNC_ENTER_NOAPI(H5S_encode, FAIL)
 
     /* Allocate "fake" file structure */
-    if(NULL == (f = H5F_fake_alloc(0)))
+    if(NULL == (f = H5F_fake_alloc((size_t)0)))
 	HGOTO_ERROR(H5E_DATASPACE, H5E_CANTALLOC, FAIL, "can't allocate fake file struct")
 
     /* Find out the size of buffer needed for extent */
